refactor: moves InputFocusWrapper to own styles dir
