+++ packages/react-tinacms-inline/src/styles/index.ts
@@ -15,11 +15,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 
 */
-
-export * from './inline-form'
-export * from './inline-field'
-export * from './inline-field-text'
-export * from './inline-field-textarea'
-export * from './inline-wysiwyg'
-export * from './blocks'
-export * from './styles'
+export * from './input-focus-wrapper'
